\name{gamselBayesVignette}
\alias{gamselBayesVignette}
\title{Display the package's vignette.}
\description{The vignette of the gamselBayes package is displayed using the default PDF file browser. It provides a detailed description of use of the package for Bayesian generalized additive model selection.}
\usage{
gamselBayesVignette()
}
\value{Nothing is returned.}
\author{Virginia X. He \email{virginia.x.he@student.uts.edu.au} and Matt P. Wand \email{matt.wand@uts.edu.au}}
\examples{\donttest{gamselBayesVignette()}
}


